+++ src/main/java/com/github/kd_gaming1/skyblockenhancements/SkyblockEnhancements.java
@@ -76,7 +76,6 @@ public class SkyblockEnhancements implements ClientModInitializer {
         HypixelLocationState.register();
         ClientPlayConnectionEvents.DISCONNECT.register((handler, client) -> {
             HypixelLocationState.reset();
-            ChatTabState.reset();
         });
 
         MissingEnchants.init();